+++ src/project/SQLiteAnalyzer.pro
@@ -27,7 +27,7 @@ SOURCES += \
     ../settings/settings.cpp \
     ../database/dbexport.cpp \
     ../database/dbschemaexport.cpp \
-    ../database/dbdataexport.cpp \
+    ../database/dbexportdata.cpp \
     ../threading/cancellation.cpp
 
 HEADERS += \
@@ -42,7 +42,7 @@ HEADERS += \
     ../settings/settings.h \
     ../database/dbexport.h \
     ../database/dbschemaexport.h \
-    ../database/dbdataexport.h \
+    ../database/dbexportdata.h \
     ../threading/cancellation.h \
     ../threading/mainthread.h
